fix: move hook logic out of interp
